get_logging_function to utils.py
