+++ tests/test_decorators.py
@@ -1,6 +1,4 @@
-import logging
 import os
-from typing import Callable
 import unittest
 
 import django
@@ -16,28 +14,12 @@ from rest_framework.response import Response
 
 import drf_logger.utils
 from drf_logger.decorators import (
-    is_request_instance, APILoggingDecorator, _get_logging_function
+    is_request_instance, APILoggingDecorator
 )
 
 factory = APIRequestFactory()
 
 logger = drf_logger.utils.get_default_logger(__name__)
-
-
-class GetLoggingFunctionTests(unittest.TestCase):
-
-    def test_simple(self):
-        """ A simple test """
-        levels = ('DEBUG', 'INFO', 'WARNING', 'ERROR', 'CRITICAL')
-        for level in levels:
-            ret_func = _get_logging_function(logger, level)
-            expected = getattr(logger, level.lower())
-            self.assertEqual(ret_func, expected)
-
-    def test_invalid_level(self):
-        """ If invalid level passed. """
-        ret_func = _get_logging_function(logger, 'invalid_level')
-        self.assertEqual(ret_func, logger.info)
 
 
 class IsRequestInstanceTests(unittest.TestCase):
